Add the command line functionality using the clingon library
Added the required code using the clingon library to implement command line functionality.
cli/options provides the command line arguments and flags for the program.
cli/command is the main function that defines the program it the command line
cli/handler is the handler function needed by the cli/command function in the clingon library
main is the entrypoint for the program